Common: Rename configuration package to agent_configuration
