+++ envs/monkey_zoo/blackbox/test_configurations/noop.py
@@ -1,4 +1,4 @@
-from common.configuration import (
+from common.agent_configuration import (
     AgentConfiguration,
     CustomPBAConfiguration,
     ExploitationConfiguration,
