+++ envs/monkey_zoo/blackbox/test_configurations/powershell_credentials_reuse.py
@@ -1,4 +1,4 @@
-from common.configuration import AgentConfiguration, PluginConfiguration
+from common.agent_configuration import AgentConfiguration, PluginConfiguration
 
 from .noop import noop_test_configuration
 from .utils import (
