+++ envs/monkey_zoo/blackbox/test_configurations/wmi_mimikatz.py
@@ -1,4 +1,4 @@
-from common.configuration import AgentConfiguration, PluginConfiguration
+from common.agent_configuration import AgentConfiguration, PluginConfiguration
 from common.credentials import Credentials, Password, Username
 
 from .noop import noop_test_configuration
