+++ monkey/infection_monkey/i_control_channel.py
@@ -1,7 +1,7 @@
 import abc
 from typing import Sequence
 
-from common.configuration import AgentConfiguration
+from common.agent_configuration import AgentConfiguration
 from common.credentials import Credentials
