+++ monkey/infection_monkey/master/exploiter.py
@@ -8,7 +8,7 @@ from threading import Event
 from typing import Callable, Dict, Sequence
 
 from common import OperatingSystems
-from common.configuration.agent_sub_configurations import (
+from common.agent_configuration.agent_sub_configurations import (
     ExploitationConfiguration,
     PluginConfiguration,
 )
