+++ monkey/monkey_island/cc/repository/i_agent_configuration_repository.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from common.configuration import AgentConfiguration
+from common.agent_configuration import AgentConfiguration
 
 
 class IAgentConfigurationRepository(ABC):
