+++ monkey/monkey_island/cc/services/telemetry/processing/state.py
@@ -1,6 +1,6 @@
 import logging
 
-from common.configuration import AgentConfiguration
+from common.agent_configuration import AgentConfiguration
 from monkey_island.cc.models import Monkey
 from monkey_island.cc.services.node import NodeService
 from monkey_island.cc.services.telemetry.zero_trust_checks.segmentation import (
