+++ monkey/tests/unit_tests/infection_monkey/master/test_exploiter.py
@@ -8,7 +8,7 @@ import pytest
 from tests.unit_tests.infection_monkey.master.mock_puppet import MockPuppet
 
 from common import OperatingSystems
-from common.configuration.agent_sub_configurations import (
+from common.agent_configuration.agent_sub_configurations import (
     ExploitationConfiguration,
     PluginConfiguration,
 )
