+++ monkey/tests/unit_tests/monkey_island/cc/repository/test_file_agent_configuration_repository.py
@@ -2,7 +2,7 @@ import pytest
 from tests.common.example_agent_configuration import AGENT_CONFIGURATION
 from tests.monkey_island import OpenErrorFileRepository, SingleFileRepository
 
-from common.configuration import AgentConfiguration
+from common.agent_configuration import AgentConfiguration
 from monkey_island.cc.repository import FileAgentConfigurationRepository, RetrievalError
